Adding the Breach Radar Algorithm
This is the breach radar algorithm as described in the paper Araujo (2017). It detects unknown common points of compromise in card fraud data, but it does not make use of their paralleledized version of the algorithm. 

Araujo, M., Almeida, M., Ferreira, J., Silva, L., & Bizarro, P. (2017, June). Breachradar: Automatic detection of points-of-compromise. In Proceedings of the 2017 SIAM International Conference on Data Mining (pp. 561-569). Society for Industrial and Applied Mathematics.