Adding a normalization module
This module gives helper functions for normalizing datasets. It provides 3 functions for normalizing a set of data, specifically, percentile normalization, which forces the data to have a uniform distribution (non-parametric), gaussian quantile normalization which forces a gaussian distribution (also non-parametric), and box-cox transformation (parametric). The box-cox transform is just a wrapper around the scipy implementation.